+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 ## stream-chat-android-state
 ### 🐞 Fixed
 - Fix thread replies updating channel unread count. [#6056](https://github.com/GetStream/stream-chat-android/pull/6056)
+- Fix messages from muted users increasing channel unread count. [#6061](https://github.com/GetStream/stream-chat-android/pull/6061)
+- Fix showing channels loading state when querying offline and no matching channels are found in the DB. [#6051](https://github.com/GetStream/stream-chat-android/pull/6051)
 
 ### ⬆️ Improved
 
@@ -70,6 +72,7 @@
 ### 🐞 Fixed
 
 ### ⬆️ Improved
+- Address some Lint findings [#6058](https://github.com/GetStream/stream-chat-android/pull/6058)
 
 ### ✅ Added